feat: add .gitignore, LICENSE, and README files; update app and graficos for improved clarity
This commit introduces a .gitignore file to exclude unnecessary files, adds a LICENSE file for project licensing, and creates a README.md for project documentation. Additionally, it updates the app.py and graficos.py files to reflect a change in terminology from "Microbiologia" to "Lodos Ativados" and enhances the clarity of the graphical interface. The requirements.txt file is also updated to include specific package versions.